Cloud Computing/Keperluan_Deploy_Model/requirements.txt
Flask==2.3.2
tensorflow==2.12.0
gunicorn
pillow
numpy
h5py
